Docs: Refactor route components
