+++ site/src/App/navigationHelpers.ts
@@ -1,10 +1,10 @@
 import groupBy from 'lodash/groupBy';
-import * as components from '../../../../lib/components';
-import { ComponentDocs } from '../../types';
-import undocumentedExports from '../../undocumentedExports.json';
+import * as components from '../../../lib/components';
+import { ComponentDocs } from '../types';
+import undocumentedExports from '../undocumentedExports.json';
 
 const componentDocsContext = require.context(
-  '../../../../lib/components',
+  '../../../lib/components',
   true,
   /.docs\.tsx$/,
 );
