+++ site/src/App/routes/foundations/layout/layout.tsx
@@ -19,14 +19,14 @@ import {
   OverflowMenu,
   MenuItem,
   TextLink,
-} from '../../../../../lib/components';
-import { HeadingProps } from '../../../../../lib/components/Heading/Heading';
-import { TextStack } from '../../TextStack/TextStack';
-import Code from '../../Code/Code';
-import tokens from '../../../../../lib/themes/wireframe/tokens';
-import { Page } from '../../../types';
+} from '../../../../../../lib/components';
+import { HeadingProps } from '../../../../../../lib/components/Heading/Heading';
+import { TextStack } from '../../../TextStack/TextStack';
+import Code from '../../../Code/Code';
+import tokens from '../../../../../../lib/themes/wireframe/tokens';
+import { Page } from '../../../../types';
 import * as styleRefs from './layout.treat';
-import { ThemedExample } from '../../ThemeSetting';
+import { ThemedExample } from '../../../ThemeSetting';
 
 const slugify = (string: string) => string.replace(/ /, '-');
 
@@ -66,7 +66,7 @@ const lipsum2 =
 
 const page: Page = {
   title: 'Layout',
-  Component: () => {
+  component: () => {
     return (
       <TextStack>
         <Heading level="2">Layout</Heading>
