+++ sku.routes.js
@@ -27,8 +27,8 @@ const getPages = relativePath => {
 const componentNames = getExports('lib/components/index.ts');
 const iconNames = getExports('lib/components/icons/index.ts');
 
-const guideRoutes = getPages('site/src/App/guides/index.ts');
-const foundationRoutes = getPages('site/src/App/foundations/index.ts');
+const guideRoutes = getPages('site/src/App/routes/guides/index.ts');
+const foundationRoutes = getPages('site/src/App/routes/foundations/index.ts');
 
 module.exports = [
   { route: '/', name: 'home' },
